Allow NMEThread.h to compile with latest version of hxcpp.  Removed some warnings.

+++ project/src/audio/WindowsAcm.cpp
@@ -185,14 +185,14 @@ public:
             {
                (WORD)WAVE_FORMAT_MPEGLAYER3,   // WORD        wFormatTag;         /* format type */
                (WORD)channels,        // WORD        nChannels;          /* number of channels (i.e. mono, stereo...) */
-               rate,       // DWORD       nSamplesPerSec;     /* sample rate */
+               (DWORD)rate,       // DWORD       nSamplesPerSec;     /* sample rate */
                128 * (1024 / 8),  // DWORD       nAvgBytesPerSec;    not really used but must be one of 64, 96, 112, 128, 160kbps
                1,        // WORD        nBlockAlign;        /* block size of data */
                0,        // WORD        wBitsPerSample;     /* number of bits per sample of mono data */
                MPEGLAYER3_WFX_EXTRA_BYTES,  // WORD        cbSize;        
             },
             MPEGLAYER3_ID_MPEG,      // WORD          wID;
-            MPEGLAYER3_FLAG_PADDING_OFF,   // DWORD         fdwFlags;
+            (DWORD)MPEGLAYER3_FLAG_PADDING_OFF,   // DWORD         fdwFlags;
             MP3_BLOCK_SIZE,       // WORD          nBlockSize;
             1,          // WORD          nFramesPerBlock;
             1393,       // WORD          nCodecDelay;
@@ -203,7 +203,7 @@ public:
          {
             WAVE_FORMAT_PCM, // WORD        wFormatTag;         /* format type */
             (WORD)channels,     // WORD        nChannels;          /* number of channels (i.e. mono, stereo...) */
-            rate,    // DWORD       nSamplesPerSec;     /* sample rate */
+            (DWORD)rate,    // DWORD       nSamplesPerSec;     /* sample rate */
             channels * rate * sizeof(short),   // DWORD       nAvgBytesPerSec;    /* for buffer estimation */
             (WORD)channels * sizeof(short),     // WORD        nBlockAlign;        /* block size of data */
             (WORD)sizeof(short)*8,     // WORD        wBitsPerSample;     /* number of bits per sample of mono data */
